Add cancelable AsyncWatch
This involved creating an asyncSendRequest on the go-etcd Client object
asyncSendRequest returns a stop channel that, when closed, attempts to cancel
the connection.

This only works if the connection is in-flight. Which means there is an
intrinsic race that cannot be avoided. If stop is called before the innards of
c.httpClient.Do(req) have reached the point where the request is registered
(and therefore cancelable) then the request will not be cancelled. For short
requests this is OK, the request is returned, and we clean up. For long-running
requests (e.g. watch) the goroutines spawned by net/http and one go-routine per
async call will leak (until the response comes in).